+++ src/adapters/node/JSONFile.ts
@@ -1,27 +1,23 @@
-import type { Adapter } from '../../core/VladikDB.js'
+import { DataFile } from './DataFile.js'
 
-import { TextFile } from './TextFile.js'
+export interface JSONFileOptions {
+  space?: number
+}
 
-export class JSONFile<T> implements Adapter<T> {
-  private adapter: TextFile
+export class JSONFile<T> extends DataFile<T> {
+  public constructor(filepath: string, options: JSONFileOptions = {}) {
+    const {
+      space = 2,
+    } = options
 
-  public constructor(path: string) {
-    this.adapter = new TextFile(path)
-  }
-
-  public async read(): Promise<T | null> {
-    const data = await this.adapter.read()
-
-    if (data === null) {
-      return null
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion
-    return JSON.parse(data) as T
-  }
-
-  public async write(data: T): Promise<void> {
-    const string = JSON.stringify(data, null, 2)
-    await this.adapter.write(string)
+    super(filepath, {
+      parse: (string) => {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-type-assertion
+        return JSON.parse(string) as T
+      },
+      stringify: (data) => {
+        return JSON.stringify(data, null, space)
+      },
+    })
   }
 }