+++ src/node.ts
@@ -1,2 +1,3 @@
+export * from './adapters/node/DataFile.js'
 export * from './adapters/node/JSONFile.js'
 export * from './adapters/node/TextFile.js'